renamed Case -> Feedback (i forgot rename controllers), better test data, embed authrority response

+++ web/grails-app/domain/web/Location.groovy
@@ -2,6 +2,6 @@ package web
 
 class Location {
 
-    Long latitude
-    Long longitude
+    Double latitude
+    Double longitude
 }
